% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee.R
\name{coffee}
\alias{coffee}
\title{A way for Sarah to limit their caffine intake}
\usage{
coffee(cups)
}
\arguments{
\item{cups}{Here sarah can enter the number of cups of coffee they have consumed}
}
\value{

}
\description{
A way for Sarah to limit their caffine intake
}
