% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fav_age.R
\name{fav_age}
\alias{fav_age}
\title{How to find someones favorite age}
\usage{
fav_age(age, birth_year, fav_age)
}
\arguments{
\item{age}{A number containing an age}

\item{birth_year}{A number that is a four diget year}

\item{fav}{A number that describes the users  favorite age}
}
\value{

}
\description{
contains one function, fav_age, that will print out peoples favorite year, age, and birth year
}
