% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lets_chat.R
\name{lets_chat}
\alias{lets_chat}
\title{A function so I can talk to Sarah ~virtually~}
\usage{
lets_chat(words)
}
\arguments{
\item{words}{Here is where the user (sarah speroff) can enter some words}
}
\value{

}
\description{
Sup dude, lets chat
}
\examples{

}
