% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bad.R
\name{theme_bad}
\alias{theme_bad}
\title{A bad example of a theme
This has very colorful lines but that's kinda it}
\usage{
theme_bad()
}
\value{

}
\description{
A bad example of a theme
This has very colorful lines but that's kinda it
}
